Fix Integers Question 5
Question 5 is mean to be an example of Integer division.

+++ src/main/java/com/ten10/training/Integers.java
@@ -13,7 +13,7 @@ public class Integers {
     double questionTwoBlank;
     int questionThreeBlank;
     int questionFourBlank;
-    double questionFiveBlank;
+    int questionFiveBlank;
     int questionSixBlank;
     int questionSevenBlank;
     String questionEightBlank;
